feat: enhance contact us and privacy policy functionalities; add role and createdBy fields to models and update related controllers

+++ src/controller/admin.controller.ts
@@ -485,7 +485,7 @@ const updateBondsmanDetails = asyncHandler(
           name,
           email,
           phoneNo,
-          countryCode
+          countryCode,
         },
       },
       {
@@ -624,38 +624,45 @@ const updateUserDetails = asyncHandler(async (req: Request, res: Response) => {
 });
 
 const createContactUs = asyncHandler(async (req: Request, res: Response) => {
-  const isAdmin = await Admin.findById(req.user?._id);
-  if (!isAdmin)
-    return res.status(403).json({ message: "only admin can allow this route" });
-
-  const { text } = req.body as {
-    text: string;
-  };
+  const { text } = req.body;
 
   if (!text || !text.trim())
-    return res.status(400).json({ message: "Text field is required" });
+    return res.status(400).json({ message: "Text is required" });
 
-  // Delete all previous entries
-  await ContactUs.deleteMany({});
+  const userId = req.user?._id;
 
+  // CHECK ADMIN
+  let role: "Admin" | "Bondsman" = "Admin";
+
+  const isAdmin = await Admin.findById(userId);
+  if (isAdmin) role = "Admin";
+  console.log({ role });
+  // Delete only same-role previous entries
+  const a = await ContactUs.deleteMany({ role });
+  console.log({ a });
   // Create new entry
   const contactUs = await ContactUs.create({
     text: text.trim(),
+    createdBy: userId,
+    role,
   });
 
-  if (!contactUs)
-    return res
-      .status(500)
-      .json({ message: "Failed to create contact us content" });
-
-  return res
-    .status(200)
-    .json({ message: "Contact us content created successfully", contactUs });
+  return res.status(200).json({
+    message: "Contact Us updated successfully",
+    contactUs,
+  });
 });
 
 const getContactUs = asyncHandler(async (req: Request, res: Response) => {
-  const contactUs = await ContactUs.findOne().sort({ createdAt: -1 });
+  const userId = req.user?._id;
 
+  // CHECK ADMIN
+  let role: "Admin" | "Bondsman" = "Admin";
+
+  const isAdmin = await Admin.findById(userId);
+  if (isAdmin) role = "Admin";
+
+  const contactUs = await ContactUs.findOne({ role });
   if (!contactUs)
     return res.status(404).json({ message: "Contact us content not found" });
 
@@ -666,25 +673,29 @@ const getContactUs = asyncHandler(async (req: Request, res: Response) => {
 
 const createPrivacyPolicy = asyncHandler(
   async (req: Request, res: Response) => {
-    const isAdmin = await Admin.findById(req.user?._id);
-    if (!isAdmin)
-      return res
-        .status(403)
-        .json({ message: "only admin can allow this route" });
-
     const { text } = req.body as {
       text: string;
     };
+    const userId = req.user?._id;
+
+    // CHECK ADMIN
+    let role: "Admin" | "Bondsman" = "Admin";
+
+    const isAdmin = await Admin.findById(userId);
+    if (isAdmin) role = "Admin";
+    console.log({ role });
+    // Delete only same-role previous entries
+    const a = await PrivacyPolicy.deleteMany({ role });
+    console.log({ a });
 
     if (!text || !text.trim())
       return res.status(400).json({ message: "Text field is required" });
 
-    // Delete all previous entries
-    await PrivacyPolicy.deleteMany({});
-
     // Create new entry
     const privacyPolicy = await PrivacyPolicy.create({
       text: text.trim(),
+      createdBy: userId,
+      role,
     });
 
     if (!privacyPolicy)
@@ -692,17 +703,23 @@ const createPrivacyPolicy = asyncHandler(
         .status(500)
         .json({ message: "Failed to create privacy policy content" });
 
-    return res
-      .status(200)
-      .json({
-        message: "Privacy policy content created successfully",
-        privacyPolicy,
-      });
+    return res.status(200).json({
+      message: "Privacy policy content created successfully",
+      privacyPolicy,
+    });
   }
 );
 
 const getPrivacyPolicy = asyncHandler(async (req: Request, res: Response) => {
-  const privacyPolicy = await PrivacyPolicy.findOne().sort({ createdAt: -1 });
+  const userId = req.user?._id;
+
+  // CHECK ADMIN
+  let role: "Admin" | "Bondsman" = "Admin";
+
+  const isAdmin = await Admin.findById(userId);
+  if (isAdmin) role = "Admin";
+  
+  const privacyPolicy = await PrivacyPolicy.findOne({ role })
 
   if (!privacyPolicy)
     return res
